Decoupling go-service into 2 parts (#6)
