+++ .gitignore
@@ -1,3 +1,4 @@
 vendor
 service
 cache
+main/main
